+++ lib/kitchen/directions/bake_index/main.rb
@@ -3,8 +3,8 @@
 module Kitchen
   module Directions
     module BakeIndex
-      def self.v1(book:)
-        V1.new.bake(book: book)
+      def self.v1(book:, types: %w[main], uuid_prefix: nil)
+        V1.new.bake(book: book, types: types, uuid_prefix: uuid_prefix)
       end
     end
   end