+++ lib/kitchen/directions/bake_index/v1.rb
@@ -65,7 +65,7 @@ module Kitchen::Directions::BakeIndex
         return -1 if force_first
         return 1 if other.force_first
 
-        name <=> other.name
+        I18n.sort_strings(name, other.name)
       end
 
       protected
@@ -112,9 +112,15 @@ module Kitchen::Directions::BakeIndex
       end
     end
 
-    def bake(book:)
+    def bake(book:, types: %w[main], uuid_prefix: '')
       @metadata_elements = book.metadata.children_to_keep.copy
-      @index = Index.new
+      @uuid_prefix = uuid_prefix
+      @indexes = types.each.with_object({}) do |type, hash|
+        index_name = type == 'main' ? 'term' : type
+        hash[index_name] = Index.new
+      end
+
+      # Numbering of IDs doesn't depend on term type
 
       book.pages.terms.each do |term_element|
         page = term_element.ancestor(:page)
@@ -132,18 +138,44 @@ module Kitchen::Directions::BakeIndex
         add_term_to_index(term_element, page_title)
       end
 
-      book.first('body').append(child: render(file: 'v1.xhtml.erb'))
+      types.each do |type|
+        @container_class = "os-index#{"-#{type}" unless type == 'main'}-container"
+        @uuid_key = "#{uuid_prefix}index#{"-#{type}" unless type == 'main'}"
+        @title = I18n.t("index.#{type}")
+
+        index_name = type == 'main' ? 'term' : type
+        @index = @indexes[index_name]
+
+        book.first('body').append(child: render(file: 'v1.xhtml.erb'))
+      end
     end
 
     def add_term_to_index(term_element, page_title)
-      group_by = I18n.transliterate(term_element.text.strip[0])
+      type =
+        if !term_element.key?('index')
+          'term'
+        elsif term_element['cxlxt:index'] == 'name'
+          'name'
+        elsif term_element['cxlxt:index'] == 'foreign'
+          'foreign'
+        end
+
+      if term_element.key?('reference')
+        term_reference = term_element['cmlnle:reference']
+        group_by = term_reference[0]
+        content = term_reference
+      else
+        group_by = I18n.transliterate(term_element.text.strip[0])
+        content = term_element.text
+      end
+
       group_by = I18n.t(:eob_index_symbols_group) unless group_by.match?(/[[:alpha:]]/)
       term_element['group-by'] = group_by
 
       # Add it to our index object
-      @index.add_term(
+      @indexes[type].add_term(
         Term.new(
-          text: term_element.text,
+          text: content,
           id: term_element.id,
           group_by: group_by,
           page_title: page_title.gsub(/\n/, '')