+++ spec/directions/bake_index/main_spec.rb
@@ -4,7 +4,7 @@ require 'spec_helper'
 
 RSpec.describe Kitchen::Directions::BakeIndex do
   it 'calls v1' do
-    expect_any_instance_of(Kitchen::Directions::BakeIndex::V1).to receive(:bake).with(book: 'blah')
-    described_class.v1(book: 'blah')
+    expect_any_instance_of(Kitchen::Directions::BakeIndex::V1).to receive(:bake).with(book: 'blah', types: %w[foo1 foo2 foo3], uuid_prefix: '')
+    described_class.v1(book: 'blah', types: %w[foo1 foo2 foo3], uuid_prefix: '')
   end
 end